requirements.txt
langgraph
langchain-core
langchain-google-genai
pydantic
python-dotenv